+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased 
 
+* [fix] More than 127 segmentation labeles allowed from Vessel Label tool
 * [new] Export to hdf5 by io3d
 * [new] Patient info in load screen by io3d
 * [new] Script for color visualization in ParaView